requirements.txt
ldap3
pysmb
python_masscan
